noisyatom/requirements.txt
appdirs==1.4.0
Django==1.10.5
django-bootstrap4==0.0.4
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-crispy-forms==1.6.1
djangorestframework==3.7.7
gunicorn==19.6.0
Markdown==2.6.9
markdown2==2.3.4
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7
Pygments==2.2.0
pyparsing==2.2.0
six==1.11.0
